refactor(query): Split large files and fix comments
- Split txn.rs (758 lines) into separate modules:
  - txn/handle.rs - TransactionHandle type
  - txn/context.rs - TransactionContext trait
  - txn/result.rs - GetTransactionResult enum
  - txn/registry.rs - TransactionRegistry trait + NoOpTransactionRegistry
  - txn/guard.rs - TransactionGuard for compile-time safety
  - txn/mod.rs - Module declarations, re-exports, and tests

- Extract DocFetcher trait to fetcher.rs from runner.rs

- Fix comment inaccuracies:
  - "warning" → "error" for TransactionGuard drop behavior
  - "NamespaceView" → "datastore" in doc_fetcher.rs

- Add tests for edge cases:
  - TransactionGuard commit failure returns error
  - Commit failure errors are retryable
  - Query errors don't invalidate transaction
