+++ crates/query/src/txn/registry.rs
@@ -1,0 +1,84 @@
+//! Transaction registry trait and no-op implementation.
+
+use async_trait::async_trait;
+
+use crate::error::TransactionError;
+
+use super::handle::TransactionHandle;
+use super::result::GetTransactionResult;
+
+/// Registry for managing active transactions.
+///
+/// The database layer implements this to track transactions that can be
+/// used by the query executor.
+#[async_trait]
+pub trait TransactionRegistry: Send + Sync {
+    /// Begin a new transaction.
+    ///
+    /// Returns a handle that can be used with `get()`, `commit()`, and `rollback()`.
+    async fn begin(
+        &self,
+        readonly: bool,
+    ) -> std::result::Result<TransactionHandle, TransactionError>;
+
+    /// Get an existing transaction by handle.
+    ///
+    /// Returns `Found(context)` if the transaction exists,
+    /// `NotFound` if it doesn't exist or has been committed/rolled back,
+    /// or `LockPoisoned` if the registry lock is poisoned.
+    fn get(&self, handle: &TransactionHandle) -> GetTransactionResult;
+
+    /// Commit a transaction.
+    ///
+    /// After commit, the handle is no longer valid for `get()`.
+    async fn commit(&self, handle: &TransactionHandle)
+        -> std::result::Result<(), TransactionError>;
+
+    /// Rollback a transaction.
+    ///
+    /// After rollback, the handle is no longer valid for `get()`.
+    async fn rollback(
+        &self,
+        handle: &TransactionHandle,
+    ) -> std::result::Result<(), TransactionError>;
+}
+
+/// A no-op transaction registry that doesn't support transactions.
+///
+/// This is used when transactions aren't needed or available.
+#[derive(Debug, Clone, Default)]
+pub struct NoOpTransactionRegistry;
+
+#[async_trait]
+impl TransactionRegistry for NoOpTransactionRegistry {
+    async fn begin(
+        &self,
+        _readonly: bool,
+    ) -> std::result::Result<TransactionHandle, TransactionError> {
+        Err(TransactionError::not_supported(
+            "transactions are not supported in this configuration",
+        ))
+    }
+
+    fn get(&self, _handle: &TransactionHandle) -> GetTransactionResult {
+        GetTransactionResult::NotFound
+    }
+
+    async fn commit(
+        &self,
+        _handle: &TransactionHandle,
+    ) -> std::result::Result<(), TransactionError> {
+        Err(TransactionError::not_supported(
+            "transactions are not supported in this configuration",
+        ))
+    }
+
+    async fn rollback(
+        &self,
+        _handle: &TransactionHandle,
+    ) -> std::result::Result<(), TransactionError> {
+        Err(TransactionError::not_supported(
+            "transactions are not supported in this configuration",
+        ))
+    }
+}
